# server/swagger-server.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Collaboratios",
    "description": "All available API for the Univeristy of Milan WebApp",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/departments": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/domainFieldSubfields": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/openAccessStatuses": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/sdgs": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/institutions": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/year": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/mapCollaborations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "institution",
            "in": "query",
            "type": "string"
          },
          {
            "name": "department",
            "in": "query",
            "type": "string"
          },
          {
            "name": "domainFieldSubfield",
            "in": "query",
            "type": "string"
          },
          {
            "name": "openAccessStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sdg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startYear",
            "in": "query",
            "type": "string"
          },
          {
            "name": "finishYear",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/institutionCollaborations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "institution",
            "in": "query",
            "type": "string"
          },
          {
            "name": "department",
            "in": "query",
            "type": "string"
          },
          {
            "name": "domainFieldSubfield",
            "in": "query",
            "type": "string"
          },
          {
            "name": "openAccessStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sdg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startYear",
            "in": "query",
            "type": "string"
          },
          {
            "name": "finishYear",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/collaborators": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "institution",
            "in": "query",
            "type": "string"
          },
          {
            "name": "department",
            "in": "query",
            "type": "string"
          },
          {
            "name": "domainFieldSubfield",
            "in": "query",
            "type": "string"
          },
          {
            "name": "openAccessStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sdg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startYear",
            "in": "query",
            "type": "string"
          },
          {
            "name": "finishYear",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/collaborations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "institution",
            "in": "query",
            "type": "string"
          },
          {
            "name": "department",
            "in": "query",
            "type": "string"
          },
          {
            "name": "domainFieldSubfield",
            "in": "query",
            "type": "string"
          },
          {
            "name": "openAccessStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sdg",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startYear",
            "in": "query",
            "type": "string"
          },
          {
            "name": "finishYear",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/authorCollaborations": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "domainFieldSubfield",
            "in": "query",
            "type": "string"
          },
          {
            "name": "openAccessStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sdg",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}